+++ ProjektAPI/Dtos/ExpensesByUserByDateResponseDto.cs
@@ -3,7 +3,7 @@
     public class ExpensesByUserByDateResponseDto
     {
         public List<ExpenseDto> Expenses { get; set; }
-        public decimal TotalExpense { get; set; }
+        public double TotalExpense { get; set; }
     }
 
 }
